# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Verisafe service API",
        "title": "Verisafe API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Open Crafts Interactive",
            "url": "https://opencrafts.io/about",
            "email": "developers@opencrafts.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.0"
    },
    "host": "qaverisafe.opencrafts.io",
    "basePath": "/",
    "paths": {
...
        "/socials/me": {
            "get": {
                "security": [
                    {
                        "BearerToken": []
                    },
                    {
                        "ApiKey": []
                    }
                ],
                "description": "Provider access and refresh tokens are never included; those are brokered through /oauth/{provider}/token instead.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "socials"
                ],
                "summary": "List the authenticated user's own linked social accounts",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/social.socialResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Missing or invalid claims",
                        "schema": {
                            "$ref": "#/definitions/core.APIError"
                        }
                    },
                    "500": {
                        "description": "Failed to fetch social connections",
                        "schema": {
                            "$ref": "#/definitions/core.APIError"
                        }
                    }
                }
            }
        },
        "/socials/user/{user_id}": {
            "get": {
                "security": [
                    {
                        "BearerToken": []
                    },
                    {
                        "ApiKey": []
                    }
                ],
                "description": "Provider access and refresh tokens are never included; those are brokered through /oauth/{provider}/token instead.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "socials"
                ],
                "summary": "List a given user's linked social accounts",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Account ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/social.socialResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid user_id",
                        "schema": {
                            "$ref": "#/definitions/core.APIError"
                        }
                    },
                    "500": {
                        "description": "Failed to fetch social connections",
                        "schema": {
                            "$ref": "#/definitions/core.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "core.APIError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "the provided input is invalid or malformed"
                }
            }
        },
...
        "pgtype.InfinityModifier": {
            "type": "integer",
            "format": "int32",
            "enum": [
                1,
                0,
                -1
            ],
            "x-enum-varnames": [
                "Infinity",
                "Finite",
                "NegativeInfinity"
            ]
        },
        "pgtype.Timestamp": {
            "type": "object",
            "properties": {
                "infinityModifier": {
                    "$ref": "#/definitions/pgtype.InfinityModifier"
                },
                "time": {
                    "description": "Time zone will be ignored when encoding to PostgreSQL.",
                    "type": "string"
                },
                "valid": {
                    "type": "boolean"
                }
            }
        },
...
        "social.socialResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "description": "Always null. See the type comment.",
                    "type": "string"
                },
                "access_token_secret": {
                    "type": "string"
                },
                "account_id": {
                    "type": "string"
                },
                "avatar_url": {
                    "type": "string"
                },
                "created_at": {
                    "$ref": "#/definitions/pgtype.Timestamp"
                },
                "description": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "expires_at": {
                    "$ref": "#/definitions/pgtype.Timestamp"
                },
                "first_name": {
                    "type": "string"
                },
                "id_token": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "nick_name": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                },
                "updated_at": {
                    "$ref": "#/definitions/pgtype.Timestamp"
                },
                "user_id": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKey": {
            "type": "apiKey",
            "name": "X-Api-Key",
            "in": "header"
        },
        "BearerToken": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}